FINAL My Answer: HP Family Tree
Exercise Problem: 
Represent information about descendant family trees from Harry Potter and design functions that operate on them.

Function 1:
Design a function that produces a pair list (i.e. list of two-element lists) of every person in the tree and his or her patrons.
Function 2:
Design a function that produces the names of every person in a given tree whose wands are made of a given material.