My Answer to: 2 One-Of P4 - Pattern Match
Exercise Problem:
Design a function that consumes Pattern and ListOf1String and produces true if the pattern matches the ListOf1String. For example,

(pattern-match? (list "A" "N" "A" "N" "A" "N")
                (list "V" "6" "T" "1" "Z" "4"))

should produce true. If the ListOf1String is longer than the pattern, but the first part matches the whole pattern produce true. If the ListOf1String is shorter than the Pattern you should produce false.