WIP My Answer to Circle Fractal
Further changes made. Very close to the expected answer of making a fractal but there is an existing gap from the main circle to the branching leaves.
This is likely due to the leaf's center not actually being in the middle of the circle. Math/geometry will be needed to adjust so that the gap is eliminated