WIP Quiz
Added a fix to Problem 2 of the quiz so that more potential "boards" (solutions/moves) of the tic tac toe board is given. 
Still need another fix so that the correct number of potential boards is generated. Right now it seems to only output one branch instead of all the branches